norme lexer
