+++ Makefile
@@ -28,6 +28,7 @@ SRC_NAME = main.c \
 		init_env.c \
 		lexer_1.c \
 		lexer_2.c \
+		lexer_3.c \
 		lst_cmd.c \
 		lst_env.c \
 		parser_1.c \
