+++ srcs/lexer_1.c
@@ -66,32 +66,8 @@ static char	*spliter(char *line, int goal)
 	return (NULL);
 }
 
-static char	*check_token(char *tk, char quote)
-{
-	char	*new_tk;
-	int		size;
-	int		i;
-	int		j;
-
-	i = -1;
-	size = 0;
-	while (tk[++i])
-		if (tk[i] == quote)
-			size++;
-	if (!(new_tk = malloc(sizeof(char) * (ft_strlen(tk) - size + 1))))
-		return (NULL);
-	i = -1;
-	j = -1;
-	while (tk[++i])
-		if (tk[i] != quote)
-			new_tk[++j] = tk[i];
-	new_tk[++j] = '\0';
-	return (new_tk);
-}
-
 static char	**check_tokens(char **tks)
 {
-	char	*new_tk;
 	int		i;
 	int		j;
 
@@ -101,19 +77,9 @@ static char	**check_tokens(char **tks)
 		j = -1;
 		while (tks[i][++j])
 		{
-			if (tks[i][j] == '\"')
+			if (tks[i][j] == '\"' || tks[i][j] == '\'')
 			{
-				new_tk = check_token(tks[i], tks[i][j]);
-				free(tks[i]);
-				tks[i] = new_tk;
-				break ;
-			}
-			else if (tks[i][j] == '\'')
-			{
-				new_tk = check_token(tks[i], tks[i][j]);
-				free(tks[i]);
-				tks[i] = change_env_to_value(new_tk);
-				free(new_tk);
+				quotes_check(tks, i, j);
 				break ;
 			}
 		}
@@ -139,32 +105,32 @@ char		**lexing(char *line)
 }
 
 /*
-int main()
-{
+   int main()
+   {
 
-	char **tk;
-	int i = 0;
-	//tk = lexing("file>tata");
-	//tk = lexing("cat salut\";o;g; ;s;h;i;t\" tout; le monde je ;vais");
-	//tk = lexing("echo dsf\"fds\"fds 	");
-	//tk = lexing("echo \"fsd   fds\" sfd 	; echo lol");
-	//tk = lexing("echo>>nana");
-	//tk = lexing("file>tata");
-	//tk = lexing("   file>>tata;  ");
-	//tk = lexing("> putain");
-	//tk = lexing("");
-	//tk = lexing("echo \"couc\"o\"\"");
-	//tk = lexing("echo \"'couc'\"o\"'\"");
-	//tk = lexing("e\"\"\"\"\"cho\" salut");
-	//tk = lexing("ifconfig -a | grep ether | sed \"s/ether //g\" | sed \"s/  //g\" | sed \"s/ //g\"");
-	tk = lexing("ifconfig -a | grep ether | sed \"s/ether //g\" | sed s/  //g\" | sed \"s/ //g\"");
-	while(tk[i])
-	{
-		printf("i=%d (%s)\n", i, tk[i]);
-		free(tk[i]);
-		i++;
-	}
-	free(tk);
-	while (1);
+   char **tk;
+   int i = 0;
+//tk = lexing("file>tata");
+//tk = lexing("cat salut\";o;g; ;s;h;i;t\" tout; le monde je ;vais");
+//tk = lexing("echo dsf\"fds\"fds 	");
+//tk = lexing("echo \"fsd   fds\" sfd 	; echo lol");
+//tk = lexing("echo>>nana");
+//tk = lexing("file>tata");
+//tk = lexing("   file>>tata;  ");
+//tk = lexing("> putain");
+//tk = lexing("");
+//tk = lexing("echo \"couc\"o\"\"");
+//tk = lexing("echo \"'couc'\"o\"'\"");
+//tk = lexing("e\"\"\"\"\"cho\" salut");
+//tk = lexing("ifconfig -a | grep ether | sed \"s/ether //g\" | sed \"s/  //g\" | sed \"s/ //g\"");
+tk = lexing("ifconfig -a | grep ether | sed \"s/ether //g\" | sed s/  //g\" | sed \"s/ //g\"");
+while(tk[i])
+{
+printf("i=%d (%s)\n", i, tk[i]);
+free(tk[i]);
+i++;
 }
-*/
+free(tk);
+while (1);
+}
+ */
